Refactored settings page & added persistant store
