+++ package.json
@@ -8,7 +8,6 @@
     "eject": "expo eject"
   },
   "dependencies": {
-    "@react-native-community/async-storage": "^1.8.0",
     "aws-amplify": "^1.2.4",
     "aws-amplify-react-native": "^2.2.3",
     "bootstrap": "^4.4.1",
